Maintain Login State on Browser Refresh
Update to maintain login state when browser window is refreshed.

User login state is saved in session storage, and session storage is checked for user login state when the browser is refreshed.

If the user login information is found in the session storage, the user is allowed to view the application.

If the user login information is not found in the session storage, the user is routed to the login screen.

Functionality implemented using signals.