Home Screen Update - Delete Home
The home screen components have been updated to allow the user to delete a home.

The delete action is cascading, which means when a home is deleted, all associated locations, devices, and temperature readings are also deleted.

When a home is deleted, a response is received from the backend application containing the ID of the deleted home, the number of locations the home had registered and the number of devices the home had registered. One home is deleted from the 'Total Homes' count on the home screen. 'Total Locations' and 'Total Devices' are updated based on the number of locations and devices removed for the deleted home.